feat(deployment): move api keys to auth module + rename
