+++ apps/api/src/auth/model-schemas/api-key/api-key.schema.ts
@@ -1,0 +1,21 @@
+import { sql } from "drizzle-orm";
+import { boolean, pgTable, timestamp, uuid, varchar } from "drizzle-orm/pg-core";
+
+import { Users } from "@src/user/model-schemas";
+
+export const ApiKeys = pgTable("api_keys", {
+  id: uuid("id")
+    .primaryKey()
+    .notNull()
+    .default(sql`uuid_generate_v4()`),
+  userId: uuid("user_id")
+    .references(() => Users.id, { onDelete: "cascade" })
+    .notNull(),
+  apiKey: varchar("api_key").notNull().unique(),
+  name: varchar("name").notNull(),
+  description: varchar("description"),
+  createdAt: timestamp("created_at").defaultNow().notNull(),
+  updatedAt: timestamp("updated_at").defaultNow().notNull(),
+  expiresAt: timestamp("expires_at"),
+  isActive: boolean("is_active").default(true).notNull()
+});
