+++ apps/api/src/auth/model-schemas/index.ts
@@ -1,0 +1,1 @@
+export * from "./api-key/api-key.schema";
