+++ apps/api/src/auth/routes/index.ts
@@ -1,1 +1,2 @@
 export * from "./send-verification-email/send-verification-email.router";
+export * from "./api-keys/api-keys.router";
