+++ apps/api/src/core/providers/postgres.provider.ts
@@ -5,6 +5,7 @@ import { migrate } from "drizzle-orm/postgres-js/migrator";
 import postgres from "postgres";
 import { container, inject } from "tsyringe";
 
+import * as authSchemas from "@src/auth/model-schemas";
 import * as billingSchemas from "@src/billing/model-schemas";
 import { config } from "@src/core/config";
 import { PostgresLoggerService } from "@src/core/services/postgres-logger/postgres-logger.service";
@@ -15,7 +16,7 @@ const logger = LoggerService.forContext("POSTGRES");
 const migrationClient = postgres(config.POSTGRES_DB_URI, { max: 1, onnotice: logger.info.bind(logger) });
 const appClient = postgres(config.POSTGRES_DB_URI, { max: config.POSTGRES_MAX_CONNECTIONS, onnotice: logger.info.bind(logger) });
 
-const schema = { ...userSchemas, ...billingSchemas, ...deploymentSchemas };
+const schema = { ...userSchemas, ...billingSchemas, ...deploymentSchemas, ...authSchemas };
 const drizzleOptions = { logger: new DefaultLogger({ writer: new PostgresLoggerService({ useFormat: config.SQL_LOG_FORMAT === "pretty" }) }), schema };
 
 const pgMigrationDatabase = drizzle(migrationClient, drizzleOptions);
