+++ apps/api/src/user/model-schemas/index.ts
@@ -1,2 +1,1 @@
 export * from "./user/user.schema";
-export * from "./user-api-key/user-api-key.schema";
